Wrong Line deleted, new "to-do" added
The only thing left to add is to finish the if statement in the for loop iterating through forms. It should find the incorrect version of the word within the text, and replace it with the correct form of the word, and then do this again for the self.words list (which is already done).
Do not delete.