Basic Follow Functions
The follow functions are much more simple and correct than the follow function implemented in Connect_N. As of yet, the link class is not being used because no charge-reduction or action potential is being used. These will be the next big change, followed by an algorithm which looks for sub-nets. These sub-nets could potentially deprecate the add_word_ref function for words that have been in the flow of the program for a while, while new words would be entirely dependent on the add_word_ref function to be accurately place in a sub-net.